Split mutation and crossover to different files
