+++ src/NaiveGAflux.jl
@@ -88,8 +88,11 @@ include("shape.jl")
 include("batchsize.jl")
 include("iteratormaps.jl")
 include("archspace.jl")
-include("mutation.jl")
-include("crossover.jl")
+include("mutation/generic.jl")
+include("mutation/graph.jl")
+include("mutation/optimizer.jl")
+include("crossover/graph.jl")
+include("crossover/optimizer.jl")
 include("candidate.jl")
 include("fitness.jl")
 include("evolve.jl")
