+++ src/mutation/optimizer.jl
@@ -1,0 +1,52 @@
+"""
+    struct OptimizerMutation{F} <: AbstractMutation{FluxOptimizer}
+    OptimizerMutation(optfun)
+    OptimizerMutation(os::Union{Tuple, <:AbstractArray})
+
+Mutatates optimizers not wrapped in `ShieldedOpt` through `optfun`.
+
+Invoked recursively for `Flux.Optimiser`s.
+"""
+struct OptimizerMutation{F} <: AbstractMutation{FluxOptimizer}
+    optfun::F
+end
+OptimizerMutation(os::Union{Tuple, <:AbstractArray}, rng=rng_default) = OptimizerMutation(o -> rand(rng, os)(learningrate(o)))
+
+"""
+    LearningRateMutation(rng=rng_default)
+
+Return an `OptimizerMutation` which mutates the learning rate of optimizers.
+"""
+LearningRateMutation(rng=rng_default) = OptimizerMutation(o -> nudgelr(o, rng))
+
+(m::OptimizerMutation)(opt::Flux.Optimiser) = Flux.Optimiser(m.(opt.os))
+(m::OptimizerMutation)(o::ShieldedOpt) = o;
+(m::OptimizerMutation)(o::FluxOptimizer) = m.optfun(o)
+
+
+nudgelr(o, rng=rng_default) = sameopt(o, nudgelr(learningrate(o), rng))
+nudgelr(lr::Number, rng=rng_default) = clamp(lr + (rand(rng) - 0.5) * lr * 0.3, 1e-6, 1.0)
+
+learningrate(o::Flux.Optimiser) = prod(learningrate.(o.os))
+learningrate(o::ShieldedOpt) = learningrate(o.opt)
+learningrate(o) = o.eta
+
+newlr(o, lrf = nudgelr) = sameopt(o, lrf(learningrate(o)))
+sameopt(o, lr) = @set o.eta = lr
+
+"""
+    AddOptimizerMutation{F} <: AbstractMutation{FluxOptimizer}
+
+Adds optimizer generated by `optgen(os)` to the set of optimizers where `os` is the existing set.
+
+An attempt to merge optimizers of the same type is made using `mergeopts`.
+"""
+struct AddOptimizerMutation{F} <: AbstractMutation{FluxOptimizer}
+    optgen::F
+end
+(m::AddOptimizerMutation)(o::ShieldedOpt) = o;
+(m::AddOptimizerMutation)(o::FluxOptimizer) = m(Flux.Optimiser([o]))
+function (m::AddOptimizerMutation)(opt::Flux.Optimiser)
+    newopt = m.optgen(opt)
+    return Flux.Optimiser(mergeopts(typeof(newopt), newopt, opt.os...))
+end
